turned long tests into mocks
